# Perf improvements for sub-expres and index-exprs