--- a/jmespath/parser.py
+++ b/jmespath/parser.py
@@ -379,8 +379,10 @@ class Parser(object):
                                     token['type'], 'Invalid token')
 
     def _match(self, token_type=None):
-        if self._current_token() == token_type:
-            self._advance()
+        # inline'd self._current_token()
+        if self._tokens[self._index]['type'] == token_type:
+            # inline'd self._advance()
+            self._index += 1
         else:
             t = self._lookahead_token(0)
             lex_position = t['start']