@danielgtaylor 2015-01-23
Aside from the question above this LGTM. :ship:-it!